updated v2 version
